Physically-based Reflectance model (#10)
* Refactor some key modules into a core module

* Initial BxDF implementation

For Specular, Diffuse, and Microfacet light transport

* WIP

* Fixed fresnel dielectric reflection

Makes specular highlights look correct

* Add metal material

Similar to plastic, but with no diffuse component and fresnel conduction

* Add metal examples

* Updated bxdf utilities

To be used for PDF calculation

* BxDF sampling for specular and microfacet modules

To be used for physically-based reflections

* BSDF sampling implementation

* Recursive specular reflection implementation

* Various fixes and optimizations

* Example specular material use

Reflections work in the simple scene! But are pretty broken at the moment

* Fix clamping bug for RGB spectrum

This seems like a huge hack, is there a better way to do this? Maybe re-implement Color as spectrum, with actual wavelengths?

* Add mirror material

* Remove material trait scene dependence

Light scattering shouldn't care about the scene

* Add glass material

* Track error separately in surface interaction

So it may be used for recursive specular reflection *and* refraction

* roughness_to_alpha function for microfacet distribution

* Coordinate system function

For creating a coordinate system from a single vector

* Various bug fixes

* Normalize background view vector on initialization

* Re-implement intersections with uvs

Then calculate normal from their cross-product

* Working scene rendering with specular reflection/transmission

Glass material works when roughness is zero

* Fix failing tests

* Fix cube refraction speckling bug

* Fix surface interaction normal again

So much normalization

* Correct specular PDFs

Should be zero

* New examples and example updates

For new reflectance model

* Fresnel -> Substance

Fresnel was a great guy but it's hard to remember what he does

* More effective microfacet transmission model

With other fixes

* Fix bxdf warnings

* Fix BSDF bug

* Fix warnings

* Fix cornell example

* New material support in lasgun-js

* Update sample JS scenes to use new materials